# src/main/java/org/teamlearning/springbootdemo/dto/ErrorMessagesDTO.java
package org.teamlearning.springbootdemo.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;


@Data
@AllArgsConstructor
@NoArgsConstructor
public class ErrorMessagesDTO {
    @JsonProperty("error")
    private String error = null;

    @JsonProperty("code")
    private String code = null;

}