# src/main/java/org/teamlearning/springbootdemo/dto/SignupDTO.java
package org.teamlearning.springbootdemo.dto;


import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;


@Data
@AllArgsConstructor
@NoArgsConstructor
public class SignupDTO {
    @JsonProperty("name")
    private String name;

    @JsonProperty("phone")
    private String phone;

    @JsonProperty("address")
    private String address;

    @JsonProperty("user")
    private String user;

    @JsonProperty("email")
    private String email;

    @JsonProperty("password")
    private String password;
}