# src/main/java/org/teamlearning/springbootdemo/dto/TaskResponseDTOs.java
package org.teamlearning.springbootdemo.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class TaskResponseDTOs {
    List<TaskResponseDTO> taskResponseDTOS;
}